requirements.txt
# AirSim Python API
msgpack-rpc-python
airsim

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# Data Processing
pandas>=2.0.0
h5py>=3.9.0
tqdm>=4.66.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Domain Randomization (if using)
gym>=0.26.0
gymnasium>=0.29.0

# Optional: For advanced ML workflows
scikit-learn>=1.3.0
scikit-image>=0.21.0

# YOLOv8 for Obstacle Detection (Phase 1)
ultralytics>=8.0.0


